pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
]
